Refactor Peer struct into peer.rs
